+++ src/django/api/tests.py
@@ -8,6 +8,7 @@ from django.test import TestCase, override_settings
 from django.urls import reverse
 from django.contrib import auth
 from django.conf import settings
+from django.contrib.auth.models import Group
 from django.contrib.gis.geos import Point
 
 from rest_framework import status
@@ -738,7 +739,26 @@ class FacilityNamesAddressesAndContributorsTest(TestCase):
         sources = self.facility.sources()
         self.assertIn(self.source_one, sources)
         self.assertNotIn(self.source_two, sources)
-        self.assertEqual(len(sources), 1)
+        self.assertIn("One Other", sources)
+        self.assertEqual(len(sources), 2)
+
+        other_names = self.facility.other_names()
+        self.assertNotIn(self.name_two, other_names)
+        self.assertEqual(len(other_names), 0)
+
+        other_addresses = self.facility.other_addresses()
+        self.assertNotIn(self.address_two, other_addresses)
+        self.assertEqual(len(other_addresses), 0)
+
+    def test_excludes_private_matches_from_details(self):
+        self.source_two.is_public = False
+        self.source_two.save()
+
+        sources = self.facility.sources()
+        self.assertIn(self.source_one, sources)
+        self.assertNotIn(self.source_two, sources)
+        self.assertIn("One Other", sources)
+        self.assertEqual(len(sources), 2)
 
         other_names = self.facility.other_names()
         self.assertNotIn(self.name_two, other_names)
@@ -4198,7 +4218,7 @@ class FacilityHistoryEndpointTest(FacilityAPITestCaseBase):
 
         self.assertEqual(
             data[0]['detail'],
-            'Associate facility {} with contributor {} via list {}'.format(
+            'Associate facility {} with {} via list {}'.format(
                 self.facility_two.id,
                 self.contributor.name,
                 self.list_two.name,
@@ -4208,6 +4228,23 @@ class FacilityHistoryEndpointTest(FacilityAPITestCaseBase):
         self.assertEqual(
             len(data),
             2,
+        )
+
+        self.superuser.groups.add(
+            Group.objects.get(name=FeatureGroups.CAN_SUBMIT_PRIVATE_FACILITY))
+        self.superuser.save()
+
+        automatic_match_response = self.client.get(
+            self.facility_two_history_url,
+        )
+
+        data = json.loads(automatic_match_response.content)
+
+        self.assertEqual(
+            data[0]['detail'],
+            'Associate facility {} with an Other'.format(
+                self.facility_two.id,
+            ),
         )
 
     @override_flag('can_get_facility_history', active=True)
@@ -4276,7 +4313,7 @@ class FacilityHistoryEndpointTest(FacilityAPITestCaseBase):
 
         self.assertEqual(
             data[0]['detail'],
-            'Associate facility {} with contributor {} via list {}'.format(
+            'Associate facility {} with {} via list {}'.format(
                 self.facility_two.id,
                 self.contributor.name,
                 self.list_for_confirm_or_remove.name,
@@ -4286,6 +4323,23 @@ class FacilityHistoryEndpointTest(FacilityAPITestCaseBase):
         self.assertEqual(
             len(data),
             3,
+        )
+
+        self.user.groups.add(
+            Group.objects.get(name=FeatureGroups.CAN_SUBMIT_PRIVATE_FACILITY))
+        self.user.save()
+
+        confirmed_match_response = self.client.get(
+            self.facility_two_history_url,
+        )
+
+        data = json.loads(confirmed_match_response.content)
+
+        self.assertEqual(
+            data[0]['detail'],
+            'Associate facility {} with an Other'.format(
+                self.facility_two.id,
+            ),
         )
 
     @override_flag('can_get_facility_history', active=True)
@@ -4344,7 +4398,7 @@ class FacilityHistoryEndpointTest(FacilityAPITestCaseBase):
 
         self.assertEqual(
             data[0]['detail'],
-            'Dissociate facility {} from contributor {} via list {}'.format(
+            'Dissociate facility {} from {} via list {}'.format(
                 self.facility_two.id,
                 self.contributor.name,
                 self.list_for_confirm_or_remove.name,
@@ -4354,6 +4408,23 @@ class FacilityHistoryEndpointTest(FacilityAPITestCaseBase):
         self.assertEqual(
             len(data),
             4,
+        )
+
+        self.user.groups.add(
+            Group.objects.get(name=FeatureGroups.CAN_SUBMIT_PRIVATE_FACILITY))
+        self.user.save()
+
+        confirmed_match_response = self.client.get(
+            self.facility_two_history_url,
+        )
+
+        data = json.loads(confirmed_match_response.content)
+
+        self.assertEqual(
+            data[0]['detail'],
+            'Dissociate facility {} from an Other'.format(
+                self.facility_two.id,
+            ),
         )
 
     @override_flag('can_get_facility_history', active=True)
@@ -4831,3 +4902,188 @@ class FacilityCreateBodySerializerTest(TestCase):
         })
         self.assertFalse(serializer.is_valid())
         self.assertIn('country', serializer.errors)
+
+
+class FacilitySearchContributorTest(FacilityAPITestCaseBase):
+    def setUp(self):
+        super(FacilitySearchContributorTest, self).setUp()
+        self.url = reverse('facility-list')
+        self.private_user = User.objects.create(email='shh@hush.com')
+        self.private_user_password = 'shhh'
+        self.private_user.set_password(self.private_user_password)
+        self.private_user.groups.set(
+            auth.models.Group.objects.filter(
+                name__in=[
+                    FeatureGroups.CAN_SUBMIT_FACILITY,
+                    FeatureGroups.CAN_SUBMIT_PRIVATE_FACILITY
+                ]).values_list('id', flat=True))
+        self.private_user.save()
+        self.client.logout()
+
+    def fetch_facility_contributors(self, facility):
+        facility_url = '{}{}/'.format(self.url, facility.id)
+        response = self.client.get(facility_url)
+        data = json.loads(response.content)
+        return data.get('properties', {}).get('contributors', [])
+
+    def test_names(self):
+        self.source.is_active = False
+        self.source.save()
+
+        self.contributor.contrib_type = 'Auditor'
+        self.contributor.save()
+        contributors = self.fetch_facility_contributors(self.facility)
+        self.assertEqual(1, len(contributors))
+        self.assertEqual('An Auditor',
+                         contributors[0].get('name'))
+
+        self.contributor.contrib_type = 'Brand/Retailer'
+        self.contributor.save()
+        contributors = self.fetch_facility_contributors(self.facility)
+        self.assertEqual(1, len(contributors))
+        self.assertEqual('A Brand/Retailer',
+                         contributors[0].get('name'))
+
+    def test_inactive_contributor(self):
+        contributors = self.fetch_facility_contributors(self.facility)
+        self.assertEqual(1, len(contributors))
+        self.assertEqual('test contributor 1 (First List)',
+                         contributors[0].get('name'))
+
+        self.source.is_active = False
+        self.source.save()
+        contributors = self.fetch_facility_contributors(self.facility)
+        self.assertEqual(1, len(contributors))
+        self.assertEqual('One Other', contributors[0].get('name'))
+
+    def test_private_contributor(self):
+        contributors = self.fetch_facility_contributors(self.facility)
+        self.assertEqual(1, len(contributors))
+        self.assertEqual('test contributor 1 (First List)',
+                         contributors[0].get('name'))
+
+        self.source.is_public = False
+        self.source.save()
+        contributors = self.fetch_facility_contributors(self.facility)
+        self.assertEqual(1, len(contributors))
+        self.assertEqual('One Other', contributors[0].get('name'))
+
+    def test_multiple(self):
+        user_two = User.objects.create(email='2@two.com')
+        user_two.set_password('shhh')
+        user_two.save()
+
+        contributor_two = Contributor \
+            .objects \
+            .create(admin=user_two,
+                    name='test contributor 2',
+                    contrib_type=Contributor.OTHER_CONTRIB_TYPE)
+
+        source_two = Source \
+            .objects \
+            .create(source_type=Source.SINGLE,
+                    is_active=True,
+                    is_public=True,
+                    contributor=contributor_two)
+
+        list_item_two = FacilityListItem \
+            .objects \
+            .create(name='Item 2',
+                    address='Address',
+                    country_code='US',
+                    row_index=0,
+                    geocoded_point=Point(0, 0),
+                    status=FacilityListItem.CONFIRMED_MATCH,
+                    source=source_two,
+                    facility=self.facility)
+
+        FacilityMatch \
+            .objects \
+            .create(status=FacilityMatch.AUTOMATIC,
+                    facility=self.facility,
+                    facility_list_item=list_item_two,
+                    confidence=0.85,
+                    results='')
+
+        source_three = Source \
+            .objects \
+            .create(source_type=Source.SINGLE,
+                    is_active=True,
+                    is_public=True)
+
+        list_item_three = FacilityListItem \
+            .objects \
+            .create(name='Item 3',
+                    address='Address',
+                    country_code='US',
+                    row_index=0,
+                    geocoded_point=Point(0, 0),
+                    status=FacilityListItem.CONFIRMED_MATCH,
+                    source=source_three,
+                    facility=self.facility)
+
+        FacilityMatch \
+            .objects \
+            .create(status=FacilityMatch.AUTOMATIC,
+                    facility=self.facility,
+                    facility_list_item=list_item_three,
+                    confidence=0.85,
+                    results='')
+
+        contributors = self.fetch_facility_contributors(self.facility)
+        self.assertEqual(2, len(contributors))
+
+        source_two.is_active = False
+        source_two.save()
+        contributors = self.fetch_facility_contributors(self.facility)
+        self.assertEqual(2, len(contributors))
+        self.assertEqual('test contributor 1 (First List)',
+                         contributors[0].get('name'))
+        self.assertEqual('One Other',
+                         contributors[1].get('name'))
+
+        self.match.is_active = False
+        self.match.save()
+        contributors = self.fetch_facility_contributors(self.facility)
+        self.assertEqual(1, len(contributors))
+        self.assertEqual('2 Others', contributors[0].get('name'))
+
+    def test_private_user(self):
+        self.client.login(email=self.private_user.email,
+                          password=self.private_user_password)
+        contributors = self.fetch_facility_contributors(self.facility)
+        self.assertEqual(1, len(contributors))
+        self.assertEqual('One Other', contributors[0].get('name'))
+
+        self.private_user.groups.set(
+            auth.models.Group.objects.filter(
+                name__in=[
+                    FeatureGroups.CAN_SUBMIT_FACILITY,
+                    FeatureGroups.CAN_SUBMIT_PRIVATE_FACILITY,
+                    FeatureGroups.CAN_VIEW_FULL_CONTRIB_DETAIL
+                ]).values_list('id', flat=True))
+        self.private_user.save()
+        contributors = self.fetch_facility_contributors(self.facility)
+        self.assertEqual(1, len(contributors))
+        self.assertEqual('test contributor 1 (First List)',
+                         contributors[0].get('name'))
+
+    def test_inactive_or_private_contributor_omitted(self):
+        def get_facility_count():
+            url = '{}?contributors={}'.format(
+                self.url, self.contributor.id)
+            response = self.client.get(url)
+            data = json.loads(response.content)
+            return int(data.get('count'))
+
+        self.assertEqual(1, get_facility_count())
+
+        self.source.is_public = False
+        self.source.is_active = True
+        self.source.save()
+        self.assertEqual(0, get_facility_count())
+
+        self.source.is_public = True
+        self.source.is_active = False
+        self.source.save()
+        self.assertEqual(0, get_facility_count())