+++ CHANGELOG.md
@@ -18,6 +18,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Security
 
+## [2.18.0] - 2019-11-12
+### Added
+
+### Changed
+- Train persistant Dedupe model on first app request [#905](https://github.com/open-apparel-registry/open-apparel-registry/pull/905/)
+- List inactive or private sources by contributor type [#907](https://github.com/open-apparel-registry/open-apparel-registry/pull/907)
+- Add can view full contributor detail flag/group [#910](https://github.com/open-apparel-registry/open-apparel-registry/pull/910)
+- Add downloadable XLSX contributor template [#912](https://github.com/open-apparel-registry/open-apparel-registry/pull/912)
+
+### Deprecated
+
+### Removed
+
+### Fixed
+- Handle EmptyResultSet error raised by tile requests [#911](https://github.com/open-apparel-registry/open-apparel-registry/pull/911)
+
+### Security
+
 ## [2.17.0] - 2019-11-04
 ### Added
 - Add single facility submission endpoint [#896](https://github.com/open-apparel-registry/open-apparel-registry/pull/896)